requirements.txt
asgiref==3.8.1
django==4.2.26
django-allauth==65.3.0
Pillow==10.4.0
gunicorn==23.0.0
packaging==25.0
python-decouple==3.8
sqlparse==0.5.3
typing-extensions==4.13.2
tzdata==2025.2
whitenoise==6.7.0
requests==2.32.4
PyJWT==2.9.0
cryptography==46.0.3

# Database
psycopg2-binary==2.9.10

# Cloud Storage
django-storages==1.14.4
google-cloud-storage==2.18.2
